% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditional_operator.R
\name{?}
\alias{?}
\title{Conditional (Ternary) Operator}
\usage{
lhs ? rhs
}
\arguments{
\item{lhs}{A logical expression, vector or matrix.}

\item{rhs}{A pair of values separated by a colon (\code{:}).}
}
\value{
One of the values in \code{rhs}, depending on the truthiness of \code{lhs}.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
Who has time for if/else?
}
\examples{
# Conditional evaluation
4 > 3 ? "it_was_true":"it_was_false"
# > "it_was_true"

FALSE ? "it_was_true":"it_was_false"
# > "it_was_false"

# Vectorised evaluation
c(4, 2) < 3 ? "it_was_true":"it_was_false"
# > "it_was_false" "it_was_true"

# Conditional assignment with `<-`
x <- 4 > 3 ? "it_was_true":"it_was_false"
x
# > "it_was_true"

# Conditional assignment with `=`
y <- 3 > 4 ? "it_was_true":"it_was_false"
y
# > "it_was_false"

# Chaining `?` statements
z <- FALSE ? "true":(FALSE ? "false,true":(TRUE ? "false,false,true":"all false"))
z
# > "false,false,true"
}
