% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditional_operator.R
\name{?}
\alias{?}
\title{Conditional (Ternary) Operator}
\usage{
`?`(lhs, rhs)
}
\arguments{
\item{lhs}{An expression that returns a logical value, or an assignment to an
expression which returns a logical.}

\item{rhs}{A pair of expressions separated by a colon (\code{:}).}
}
\description{
Who has time for if/else?
}
