% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditional_operator.R
\name{?}
\alias{?}
\title{Conditional (Ternary) Operator}
\usage{
lhs ? rhs
}
\arguments{
\item{lhs}{A logical expression  or logical vector.}

\item{rhs}{Two values separated by a colon (\code{:}).}
}
\value{
One of the values in \code{rhs}, depending on truthiness of \code{lhs}.
}
\description{
Who has time for if/else?
}
\examples{
# Conditional evaluation
4 > 3 ? "it_was_true" : "it_was_false"
#> "it_was_true"

FALSE ? "it_was_true" : "it_was_false"
#> "it_was_false"

# Vectorised evaluation
c(4, 2) < 3 ? "it_was_true" : "it_was_false"
#> "it_was_false" "it_was_true"

# Conditional assignment
x <- 4 > 3 ? "it_was_true" : "it_was_false"
x
#> "it_was_true"

y = 3 > 4 ? "it_was_true" : "it_was_false"
y
#> "it_was_false"
}
