% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_assignment.R
\name{is_assignment}
\alias{is_assignment}
\title{Check if an expression is an assignment}
\usage{
is_assignment(parse_tree)
}
\arguments{
\item{parse_tree}{An unevaluated expression as a list}
}
\value{
A logical
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
Takes an expression that has been cast to a list and returns true if using
left-assignment with either \verb{<-} or \code{=}
}
\keyword{internal}
